Extract UI static route helper
